ファイルディレクトリ修正、ビルド導通
