+++ ohgiri-app/View/AllTableViewCell.swift
