+++ ohgiri-app/View/GroupTableViewCell.swift
