+++ ohgiri-app/View/GroupTableViewCell.xib
