+++ ohgiri-app/View/PostedTableViewCell.swift
