+++ ohgiri-app/View/PostedTableViewCell.xib
